tools/aisearch_ingestion/pyproject.toml
[project]
name = "aisearch-ingestion"
version = "0.1.0"
description = "Azure AI Search index creation and CSV ingestion with vector embeddings."
requires-python = ">=3.11"
dependencies = [
    "azure-search-documents>=11.6.0",
    "azure-identity",
    "openai",
    "python-dotenv",
]

[tool.uv]
package = false
